Move non-Basic tests to their own folder [API-386]
