+++ phpunit.xml
@@ -13,6 +13,9 @@
         <testsuite name="Contract">
             <directory suffix="Test.php">./tests/Contract</directory>
         </testsuite>
+        <testsuite name="Unit">
+            <directory suffix="Test.php">./tests/Unit</directory>
+        </testsuite>
     </testsuites>
     <coverage>
         <include>
